requirement.txt
discord
python-dotenv
selenium
json
webdriver-manager
chromedriver_binary
yaml
getpass
youtube_dl
PyNaCl
ffmpeg